move view_boundaries to src/idfhub/helpers/geometry.py
